WIP: Basic input form.
(Almost) all the rows for the input form, with classes added from the
timetable — the ‘almost’ being because I haven't added the principal roles
yet.

JQuery to duplicate the widgets across 4 columns, tweaking the names and
classes accordingly, and setting up the Tab order to go down the columns,
then focus the first control.