Fix: Actually unique tab order
Within each column, each widget should have its own tab index. This was
always the intention (hence leaving a gap of 10 per column), but I took it
out at some point to eliminate it from inquiries when debugging, and
forgot to put it back once it turned out to be blameless.

So here it is.

It was happening to work anyway, presumably because equal-tab-indexed
elements are navigated in source order. But since we have actual unique
numbers available, there's no reason to rely on that.