Progress: Act 3 rehearsal data
Added who's required in Act 3.

Also Adult Bollywood exists, and tidying up some other bits.

Peter and Susan need separate letters, because they are separate items in
the list and when saving is introduced we need to be able to set to the
correct item on loading. Similarly, Mr & Mrs Beaver (who I think might
have different timings for one of the in-theatre rehearsals, because of
the one dance that Mr Beaver is in without Mrs Beaver).

‘mb’ has been retconned to ‘male beaver’, so ‘fb’ can be used for Mrs
Beaver.