Enhancement: Print button
Have a ‘Print’ button.

Make it hidden by default, displayed when a schedule is created.

Have it in grey, rather than the green of the ‘Create Schedule’ button.
Or, rather, have all buttons default to grey and make it that the go
button is special by being green. Add a bit of space between the buttons.