Refactoring: Letter codes for non-numeric classes
Use code 'ss' for Stepping Stones (instead of -2). This is way clearer for
the places where we specifically want Stepping Stones.

And it's also fine for grouping classes -1 to 1 numerically: 'ss' <= 1
returns false, because it isn't a number, so 'ss' doesn't get included
anyway, and we don't even need the >= -1 condition any more. And similarly
for the >= 2 for costume collection.

On that basis, also change the adult tap codes, since the numbers were so
arbitrary, to the more mnemonic 'ab' and 'ai'.