Enhancement: Save feature, using a cookie
Instead of id=print have class=action, for hiding all action buttons that
only make sense once a schedule is displayed.

When generating a schedule, have the variable the state is saved in be
global, so it's still available if ‘Save’ is later pressed.

To show that saving has done something, disable the save button on saving.
Undisable all action buttons when generating a schedule, so make save
usable again.

So that that's visible straight away, have the :hover highlight effect
only apply when the button isn't disabled.

When loading the page, if there isn't any state in the URL then look for
it in a cookie instead — a pleasingly small change to add this
functionality, and have it just work!

Use the js-cookie library, because cookies are tedious to parse:
https://github.com/js-cookie/js-cookie