Enhancement: Adult Trees
Gemma says adult trees are a subset of the Bollywood class. Since there's
only one Bollywood class, and it doesn't have a time on it, pretend it's a
separate class, called ‘Adult Bollywood + Tree’.

Everywhere that's currently testing for Bollywood is doing so as a
boolean, so those will all continue to be true.

Add the trees to the scenes that require them.