Enhancement: Intermediate Ballet
From Gemma: The Intermediate ballet class is what's on the timetable as
Grade 6.

(That actually works better, given I've assigned them a value of 7!)

They are in 2 dances, both with Class Level 6 (who have a value of 6),
which makes it easy to compare them together. One already was doing that.