Refactoring: Fonts loaded in CSS
Avoid the action-at-a-distance of specifying the font families in the CSS
file but loading the files for them in the HTML by also doing the loading
the CSS.

This also has less repetition, since it only needs specifying in one CSS
file rather than every HTML file.

The preconnects still have to be in the HTML. But those are presumably
just for efficiency, so things would still work without them, and they are
generic to the Google Fonts service, so not as tightly coupled as
specifying individual font names.

Finally, this also avoids an HTML spec violation, where the ampersands in
Google's URLs weren't being escaped as &amp;