Fix: Class Level 4 are actually 5
The ballet class that wasn't included in any dances is actually now the
next class level up. Unfortunately this gives them the same name as the
next class up, and I don't know what that should be.

Logic says to put them up a level as well, but the combination of levels
and grades makes that non-obvious. And there's still one more class in the
timetable then the schedule, so if every class has a unique name then one
of them isn't in the show!

For now leave them with ID 4. Once I have actual names for the classes
above them, I can make the IDs match.

But do add them to the dances that are labelled for Class Level 5, by
including ID 4 in them. This makes the ranges no longer have a gap in
them, which seems much more sensible.