Fix: All ballet classes now correct
From Nikki, Class Level 5 ballet did the exam and are now Class Level 6.
But the existing Class Level 6 class *didn't* do the exam, and are *still*
Class Level 6 as well.

So the only remaining fix is renaming the Friday Class Level 5 Class.

I wanted to tweak the IDs. (Having 2 classes at the same level means
somebody is going to be non-intuitive, but there's a split between classes
at up to Level 5 and Level 6 upwards, so those are the ones it seems most
important to be as expected.)

But because users may well have bookmarked or saved state with IDs in now,
it would terrible if it suddenly changed which classes they are in. So
keep the IDs as they were and put up with dances up to Class Level 5 being
<= 4 and those from Class Level 6 upwards being >= 5. Bah.