Fix: Babies of the Blitz tap levels
Nikki has clarified who is in the Babies of the Blitz scene. It's modern
classes 5 and 6.

Lynne had used tap grades for these, but modern grades for the dance below
— leaving modern grade 4 in both dances and modern grade 6 in neither.

Now that this dance is grade 5+ plus adults, having the adults as IDs 10
and 11 like I originally did would be helpful here, and avoid needing an
error-prone .includes check: it could just be dancer.modern >= 5.

However, like with the previous commit, it could be awkward to change the
IDs now, so let's stick with it for this year.