Fix: Adult costume collection
(That's costume collection for those in the Adult Beginners and Adult
Improvers classes. Not collection of ‘adult’ costumes.)

Modern/tap costumes are being collected for grades >= 2. Which is supposed
to include the adults, but doesn't because 'ab' and 'ai' are not >= 2, not
being numeric at all.

So, despite what I said in the parent commit, *do* change the adult tap
classes back to numeric IDs, reversing the change in d64c6e87.

The ID change is unfortunate but doesn't matter as much here for 2
reasons. The main one is that the old IDs (the ones with letters) aren't
being re-used as new IDs, so the form simply forgets which class you were
in, rather than subtly changing you to a different (but similar) class.

The other reason is that adult tap are going to be in so few classes
(possibly indeed only that one) that it's going to be obvious when a class
is missing.

This also allows simplifying the Babies of the Blitz cast.

Handily the other places which are grades 5 and 6 already have an upper
limit of 6 on anyway — which was unnecessary until now, but is now needed
for excluding the adults.

However, there was one dance listed as modern grades 4+, coded as simply
>= 4. Given adults don't do modern, that's unambiguous and reasonable in
the schedule, but it does now need capping at 6 to exclude the adults.