Fix: Grade 2 Tap/Modern not collecting costumes
Costume collection was announced as being for Grade 2+ — and specifically
to exclude the Wednesday 16:00 class, but not to exclude (as in, they
still need to collect) the Wednesday 17:00 class.

But I can't see them anywhere else in the schedule, and Nikki has
confirmed: “The grade 2 modern and tap are only evacuees. If they are not
in any other style like Ballet they don't need to collect costumes as they
are sourcing their own.”

So let's not unnecessarily send them to Ben Rhydding tomorrow.