Enhancement: Times depend on gender
Specify whether male or female, and allow times to depend on those.

Girl snowflakes need to stay an extra half-hour this morning for tutu
fitting.

(There will also be differences in show arrival times, but this is
sufficient both to require the feature and demonstrate that it works.)

Have female be the default gender, because that applies to most dancers,
so this is the least surprising to appear when somebody loads a saved or
bookmarked schedule.

And being the default have it use "" as its value. This simplifies several
things compared with using the more obvious "f":

• Determining how many output columns to display involves counting how
  many input columns have any field set to something other than "". So
  using "f" for female would need an exception to avoid all 4 columns
  always being displayed.

• Cloning columns and clearing the form sets all values to "". That
  doesn't work for a <select> which doesn't have such an option; it sets
  it to a blank label that doesn't appear anywhere. So that would need
  exceptions as well.

Having f as the default using the default value of "" just avoids needing
any exceptions.